Add files via upload
I am a novice programmer teaching himself python. I have some experience with C language in a college classroom. This game of black jack was started as a way of comparing apples to apples, I made a blackjack game in c and wanted to explore doing the exact same thing in Python. I will continue to improve this game as I have time and I plan on making an object oriented version in the future as I gain a better grasp on the Python language.